lightrag/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.12.0",
    "gradio>=5.49.0",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "langchain-huggingface>=0.3.1",
    "langchain-text-splitters>=0.3.11",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.8",
    "pymupdf>=1.26.4",
    "networkx>=3.0",
    "python-dotenv>=1.0.0",
]
